% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{count_reactions}
\alias{count_reactions}
\title{count_reactions
Count the occurrences of each event (at any meddra level)
and \% as the number of individuals in the sample recording the event.}
\usage{
count_reactions(pids_cases, meddra_level = "pt")
}
\arguments{
\item{pids_cases}{the primaryids identifying the sample}

\item{meddra_level}{meddra_level investigated}
}
\value{
a database with three columns: event, N, and \% in descending order.
}
\description{
count_reactions
Count the occurrences of each event (at any meddra level)
and \% as the number of individuals in the sample recording the event.
}
\examples{
\dontrun{
count_reactions(pids_cases, "pt")
}
}
