% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_dictionary}
\alias{country_dictionary}
\title{Country Dictionary Dataset}
\format{
A data frame with 3 variables:
\describe{
\item{countryname}{Character. The name of the country, which can include official names and various abbreviations.}
\item{country}{Character. The official name of the country.}
\item{continent}{Character. The continent to which the country belongs.}
}
}
\usage{
country_dictionary
}
\description{
A dataset containing mappings of country names, their abbreviations, and their respective continents.
}
\details{
This dataset provides a comprehensive mapping of country names and their respective continents. It includes various forms of country names such as official names and abbreviations. It is used internally for functions such as descriptive.
}
\keyword{datasets}
