% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANA_reference.R
\name{DiAna_reference}
\alias{DiAna_reference}
\title{Get DiAna reference}
\usage{
DiAna_reference(print = TRUE)
}
\arguments{
\item{print}{Logical: print the text? Default \code{TRUE}. The text is also returned, invisibly when printed.}
}
\value{
DiAna reference
}
\description{
This function provides the DiAna reference, that should be used when using DiAna.
}
\examples{
DiAna_reference()

}
