% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{disproportionality_analisis}
\alias{disproportionality_analisis}
\title{Perform Disproportionality Analysis}
\usage{
disproportionality_analisis(
  drug_selected,
  reac_selected,
  temp_d = Drug,
  temp_r = Reac,
  meddra_level = "pt",
  restriction = "none",
  sign_ciphers = 2,
  threshold_ROR = 3,
  threshold_IC = 1,
  event_names = "none",
  KeepLessThanThreshold = TRUE,
  order = "reac",
  threshold = 1
)
}
\arguments{
\item{drug_selected}{A list of drugs for analysis.}

\item{reac_selected}{A list of adverse events for analysis.}

\item{temp_d}{Data table containing drug data (default is Drug).}

\item{temp_r}{Data table containing reaction data (default is Reac).}

\item{meddra_level}{The desired MedDRA level for analysis (default is "pt").}

\item{restriction}{Primary IDs to consider for analysis (default is "none").}

\item{sign_ciphers}{Number of decimal places for calculated values (default is 2).}

\item{threshold_ROR}{Threshold for calculating Risk Odds Ratio (default is 3).}

\item{threshold_IC}{Threshold for calculating Information Component (default is 1).}

\item{event_names}{Event names for labeling results (default is "none").}

\item{KeepLessThanThreshold}{Keep values less than threshold (default is TRUE).}

\item{order}{Order for analysis: "reac" or "drug" (default is "reac").}

\item{threshold}{Threshold for analysis (default is 1).}
}
\value{
A data.table containing disproportionality analysis results.
}
\description{
performs disproportionality analysis and returns the results.
}
