% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANA_reference.R
\name{FAERS_quarter_specifics}
\alias{FAERS_quarter_specifics}
\title{Get specifics of dictionaries used in database version}
\usage{
FAERS_quarter_specifics(quarter = FAERS_version, print = TRUE)
}
\arguments{
\item{quarter}{The quarter for which specifics are needed. Default to FAERS_version.}

\item{print}{Logical: print the text? Default \code{TRUE}. The text is also returned, invisibly when printed.}
}
\value{
Specifics
}
\description{
This function provides specifics on dictionaries used for cleaning the specified database version
}
\examples{
FAERS_quarter_specifics("24Q1")
}
