% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{fisher.test.simulate.p.values}
\alias{fisher.test.simulate.p.values}
\title{Compute Fisher's Exact Test with Simulated p-values}
\usage{
fisher.test.simulate.p.values(data, variable, by, ...)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{variable}{The name of the first categorical variable.}

\item{by}{The name of the second categorical variable.}

\item{...}{Additional arguments to be passed to the \code{\link[stats:fisher.test]{fisher.test}} function.}
}
\value{
A list containing the following elements:
\describe{
\item{p}{The simulated p-value for the Fisher's Exact Test.}
\item{test}{The name of the test method, which is "Fisher's Exact Test with Simulation".}
}
}
\description{
This function performs Fisher's Exact Test with simulated p-values for association between two categorical variables.
}
\seealso{
\code{\link[stats:fisher.test]{fisher.test}} for more information on Fisher's Exact Test.
}
