% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{format_input_disproportionality}
\alias{format_input_disproportionality}
\title{Format Input for Disproportionality Analysis}
\usage{
format_input_disproportionality(input)
}
\arguments{
\item{input}{A list or an object that can be converted to a list (specifically, drug_selected and reac_selected).}
}
\value{
A formatted list of drugs or events suitable for disproportionality analysis
}
\description{
This function formats the input of drug and reac selected for disproportionality analysis.
It ensures the input is in the correct list format and processes it
to meet specific structural requirements, allowing the researcher for a greater freedom in calling the function.
}
\examples{
input <- format_input_disproportionality(c("TGA" = list(
  "aripiprazole",
  "brexpiprazole",
  "cariprazine"
)))

}
