% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{hierarchical_rates}
\alias{hierarchical_rates}
\title{Generate Hierarchy of events or substances}
\usage{
hierarchical_rates(
  pids_cases,
  entity = "reaction",
  file_name = paste0(project_path, "reporting_rates.xlsx"),
  drug_role = c("PS", "SS", "I", "C"),
  save_in_excel = !missing(file_name),
  temp_meddra = NULL,
  temp_atc = NULL,
  ...
)
}
\arguments{
\item{pids_cases}{Vector of primary IDs identifying the sample of interest.}

\item{entity}{Entity investigated. It can be one of the following:
\itemize{
\item \emph{reaction};
\item \emph{indication};
\item \emph{substance}.
}}

\item{file_name}{Path of the XLSX file to save the hierarchy in, if \code{save_in_excel = TRUE}. Default "reporting_rates.xlsx" in \code{project_path}.}

\item{drug_role}{If entity is substance, it is possible to specify the drug roles that should be considered}

\item{save_in_excel}{Whether to also save the hierarchy in an Excel file, \code{file_name}. Defaults to \code{TRUE} only if \code{file_name} is supplied.}

\item{temp_meddra}{MedDRA dictionary, used for the levels "hlt", "hlgt" and "soc".
By default, \code{MedDRA} from your workspace if it is loaded, otherwise it is
read with \code{\link[=import_MedDRA]{import_MedDRA()}}.}

\item{temp_atc}{ATC classification, used for the levels "Class1" to "Class4".
By default, \code{ATC} from your workspace if it is loaded, otherwise it is
read with \code{\link[=import_ATC]{import_ATC()}}. Only primary ATC codes are used.}

\item{...}{Case tables passed on to \code{\link[=reporting_rates]{reporting_rates()}}: \code{temp_reac},
\code{temp_indi} and \code{temp_drug}. By default, \code{Reac}, \code{Indi} and \code{Drug} from
your workspace.}
}
\value{
A data.table with the hierarchy of interest.
For indications and reactions, SOCs are ordered by occurrences and, within, HLGTs, HLTs, PTs.
For substances, the ATC hierarchy is followed.
}
\description{
This function generates a hierarchy of reporting rates
for a specified entity and based on different MedDRA or ATC levels
and writes the result to a xlsx file.
}
\examples{
# The following examples require the MedDRA and the ATC to be available
pids <- sample_Demo$primaryid
if (file.exists(file.path(here::here(), "external_sources", "meddra_primary.csv"))) {
  MedDRA <- import_MedDRA()
  hierarchical_rates(pids, "reaction", temp_meddra = MedDRA, temp_reac = sample_Reac)
  hierarchical_rates(pids, "indication", temp_meddra = MedDRA, temp_indi = sample_Indi)
}
if (file.exists(file.path(here::here(), "external_sources", "ATC_DiAna.csv"))) {
  ATC <- import_ATC()
  hierarchical_rates(pids, "substance", temp_atc = ATC, temp_drug = sample_Drug)
}
}
