% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{hierarchy_reactions}
\alias{hierarchy_reactions}
\title{hierarchy_reactions
Count the occurrences of each event (at all the meddra level) and provides
a hierarchical csv.}
\usage{
hierarchy_reactions(pids_cases, reactions_path)
}
\arguments{
\item{pids_cases}{the primaryids identifying the sample}

\item{reactions_path}{the path for storing the csv.}
}
\value{
csv with SOCs ordered by occurrences and, within, HLGTs, HLTs, PTs.
}
\description{
hierarchy_reactions
Count the occurrences of each event (at all the meddra level) and provides
a hierarchical csv.
}
\examples{
\dontrun{
hierarchy_reactions(pids_cases, "reactions_list.csv")
}
}
