% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{network_analysis}
\alias{network_analysis}
\title{Perform Network Analysis and Visualization}
\usage{
network_analysis(
  pids,
  entity = "reaction",
  remove_singlet = TRUE,
  remove_negative_edges = TRUE,
  file_name = paste0(project_path, "network.tiff"),
  width = 1500,
  height = 1500,
  labs_size = 1,
  min_frequency_term = 0.01,
  restriction = "none",
  temp_reac = Reac,
  temp_indi = Indi,
  temp_drug = Drug,
  save_plot = !missing(file_name)
)
}
\arguments{
\item{pids}{Numeric vector of unique identifiers on which the network analysis should be performed.}

\item{entity}{Character specifying the type of entity to analyze ("reaction", "indication", or "substance").}

\item{remove_singlet}{Logical indicating whether to remove singleton nodes (nodes with no edges). Default is TRUE.}

\item{remove_negative_edges}{Logical indicating whether to remove edges with negative weights. Default is TRUE.}

\item{file_name}{Character string specifying the file name (including path) to save the network visualization, if \code{save_plot = TRUE}. Default is "network.tiff" in \code{project_path}.}

\item{width}{Numeric specifying the width of the saved image in pixels. Default is 1500.}

\item{height}{Numeric specifying the height of the saved image in pixels. Default is 1500.}

\item{labs_size}{Size of labels in network visualization. Default is 1. It can be changed if visualization is not good.}

\item{min_frequency_term}{Frequency threshold for a term in the dataset to be included in the analysis. Default to 0.01}

\item{restriction}{Restriction performed in the analysis. Default is none. It could be set to 'suspects' if entity is 'substance' to restrict the analysis to primary and secondary suspects}

\item{temp_reac}{Reac dataset. Can be set to sample_Reac for testing}

\item{temp_indi}{Indi dataset. Can be set to sample_Indi for testing}

\item{temp_drug}{Drug dataset. Can be set to sample_Drug for testing}

\item{save_plot}{Whether the plot should also be saved as a TIFF file, \code{file_name}. Defaults to \code{TRUE} only if \code{file_name} is supplied.}
}
\value{
The network, as an igraph object, which can be drawn with \code{plot()}. If \code{save_plot = TRUE}, the visualization is also saved as a TIFF file.
}
\description{
This function performs network analysis on the provided data and generates a visualization of the network. It uses IsingFit to model the network structure, clusters nodes using Louvain method, and visualizes the resulting network graph.
}
\examples{
# Perform network analysis for reactions with specified pids
net_plot <- network_analysis(
  pids = sample_Demo$primaryid,
  entity = "reaction", temp_reac = sample_Reac,
  save_plot = FALSE
)
plot(net_plot)

}
\references{
Fusaroli M, Raschi E, Gatti M, De Ponti F, Poluzzi E. Development of a Network-Based Signal Detection Tool: The COVID-19 Adversome in the FDA Adverse Event Reporting System. Front Pharmacol. 2021 Dec 8;12:740707. doi: 10.3389/fphar.2021.740707. PMID: 34955821; PMCID: PMC8694570.

Fusaroli, M., Polizzi, S., Menestrina, L. et al. Unveiling the Burden of Drug-Induced Impulsivity: A Network Analysis of the FDA Adverse Event Reporting System. Drug Saf (2024). https://doi.org/10.1007/s40264-024-01471-z
}
