% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_onset_analysis.R
\name{render_tto}
\alias{render_tto}
\title{TTO Render Forest Plot}
\usage{
render_tto(
  df,
  row = "substance",
  levs_row = NA,
  nested = FALSE,
  show_legend = TRUE,
  transformation = "log10",
  text_size_legend = 15,
  dodge = 0.3,
  nested_colors = NA,
  facet_v = NA,
  facet_h = NA
)
}
\arguments{
\item{df}{Data.table containing the data for rendering the forest plot.}

\item{row}{Variable for the rows of the forest plot (default is "substance").}

\item{levs_row}{Levels for the rows of the forest plot.}

\item{nested}{Variable indicating if nested plotting is required (default is FALSE). If nested plotting is required the name of the variable should replace FALSE.}

\item{show_legend}{Logical indicating whether to show the legend (default is FALSE).}

\item{transformation}{Transformation for the x-axis (default is "identity").}

\item{text_size_legend}{Size of text in the legend (default is 15).}

\item{dodge}{Position adjustment for dodging (default is 0.3).}

\item{nested_colors}{Vector of colors for plot elements.}

\item{facet_v}{Variable for vertical facetting (default is "NA", it could be set to e.g., "event").}

\item{facet_h}{Variable for horizontal facetting (default is NA).}
}
\value{
A ggplot object representing the forest plot visualization.
}
\description{
This function generates a forest plot visualization of time to onset analysis
}
\examples{
df <- time_to_onset_analysis(
  drug_selected = "skin care",
  reac_selected = list("skin affection" = list(
    "dry skin",
    "skin burning sensation",
    "skin irritation",
    "erythema",
    "rash macular",
    "acne",
    "skin haemorrhage"
  )),
  temp_drug = sample_Drug, temp_reac = sample_Reac,
  temp_ther = sample_Ther
)
render_tto(df)

}
