% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Drug}
\alias{sample_Drug}
\title{Sample Data on Drug Administration}
\format{
A data.table with 4 variables:
\describe{
\item{primaryid}{Unique identifier for each individual (\code{numeric}).}
\item{drug_seq}{Sequence number identifying the specific drug in the report (\code{integer}).}
\item{substance}{Name of the drug substance (active ingredient), as standardized using DiAna_Dictionary (\code{factor}).}
\item{role_cod}{Role code indicating the role of the drug (\code{ordered}), with levels \code{PS} for primary suspect, \code{SS} for secondary suspect, \code{C} for concomitant, \code{I} for interacting.}
}
}
\source{
https://github.com/fusarolimichele/DiAna_cleaning
}
\usage{
sample_Drug
}
\description{
A dataset containing information on drugs administered to individuals,
including substance names, and roles.
}
\details{
This subset of Drug provides detailed information on drugs administered to individuals,
specifying their suspected roles in the context of adverse drug reactions.
}
\keyword{datasets}
