% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Drug_Name}
\alias{sample_Drug_Name}
\title{Sample Data on Drug Names before standardization}
\format{
A data.table with 5 variables:
\describe{
\item{primaryid}{Unique identifier for each report (\code{numeric}).}
\item{drug_seq}{Sequence number identifying the specific drug in the report (\code{integer}).}
\item{val_vbm}{Validity flag indicating whether the drug name is valid or a free text (not reliable) (\code{integer}).}
\item{nda_num}{New Drug Application number (\code{character}).}
\item{drugname}{Name of the drug as recorded by the reporter (\code{factor}).}
\item{prod_ai}{Product active ingredient as recorded by the reporter (\code{factor}).}
}
}
\source{
https://github.com/fusarolimichele/DiAna_cleaning
}
\usage{
sample_Drug_Name
}
\description{
A dataset containing information on drug names and associated details,
including drug names, and additional attributes.
}
\details{
This subset of Drug_Name provides detailed information on drug names and associated attributes.
}
\keyword{datasets}
