% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{tailor_disproportionality_threshold}
\alias{tailor_disproportionality_threshold}
\title{Tailor Disproportionality Threshold}
\usage{
tailor_disproportionality_threshold(
  disproportionality_df,
  minimum_cases = 3,
  log2_threshold = 0,
  frequentist_threshold = 1,
  multiple_comparison = TRUE
)
}
\arguments{
\item{disproportionality_df}{A data frame containing disproportionality metrics. Generated by disproportionality_analysis.}

\item{minimum_cases}{An integer specifying the minimum number of cases required to consider the signal. Defaults to 3.}

\item{log2_threshold}{A numeric value specifying the threshold for the Information Component (IC) signal. Defaults to 0.}

\item{frequentist_threshold}{A numeric value specifying the threshold for the Reporting Odds Ratio (ROR) signal. Defaults to 1.}

\item{multiple_comparison}{A logical value indicating whether to apply multiple comparison adjustments (e.g., Bonferroni correction). Defaults to TRUE.}
}
\value{
A data frame with two new columns: \code{ROR_signal} and \code{IC_signal}, categorizing the signals based on the provided thresholds.
}
\description{
This function analyzes a data frame of disproportionality metrics and categorizes signals based on specified thresholds.
}
\details{
The function categorizes signals into four levels:
\itemize{
\item "not enough cases": when the number of cases is less than \code{minimum_cases}.
\item "no SDR": when the signal does not meet the specified threshold.
\item "weak SDR": when the signal disappear under multiple comparison adjustments.
\item "SDR": when the signal remain after multiple comparison.
}
}
