% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{count_reactions}
\alias{count_reactions}
\title{Count Reactions by MedDRA Level}
\usage{
count_reactions(pids_cases, meddra_level = "pt")
}
\arguments{
\item{pids_cases}{Vector of primary IDs to consider for counting reactions.}

\item{meddra_level}{The desired MedDRA level for counting (default is "pt").}
}
\value{
A data.table containing counts and percentages of reactions
at the specified MedDRA level and in descending order.
}
\description{
counts the occurrences of reactions based on MedDRA levels
for a given set of primary IDs.
Calculates \% as the proportion of individuals recording the event.
}
\examples{
\dontrun{
count_reactions(pids_cases, "pt")
}
}
