% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptive}
\alias{descriptive}
\title{Generate Descriptive Statistics for a Sample}
\usage{
descriptive(
  pids_cases,
  RG = NULL,
  drug = NULL,
  file_name = "Descriptives.xlsx",
  vars = c("sex", "Submission", "Reporter", "age_range", "Outcome", "country",
    "continent", "age_in_years", "wt_in_kgs", "Reactions", "Indications", "Substances",
    "year", "role_cod", "time_to_onset"),
  list_pids = list(),
  method = "independence_test",
  quarter = FAERS_version
)
}
\arguments{
\item{pids_cases}{A vector of primary IDs for cases.}

\item{RG}{A vector of primary IDs: reference group. Default is NULL.}

\item{drug}{A vector of drug names. Default is NULL.}

\item{file_name}{The name of the Excel file to save the results. Default is "Descriptives.xlsx".}

\item{vars}{A character vector of variable names to include in the analysis.}

\item{list_pids}{A list of vectors with primary IDs for custom groups whose distribution should be described. Default is an empty list.}

\item{method}{The method for Chi-square test analysis, either "independence_test" or "goodness_of_fit". Default is "independence_test". It applies only for comparisons between cases and non-cases.}

\item{quarter}{The quarter for data import. Default is "23Q1".}
}
\value{
The function generates descriptive statistics and saves them to an Excel file. It also creates CSV files for substances, indications, and reactions if specified.
}
\description{
This function generates descriptive statistics for a sample of data, including cases and non-cases, and saves the results to an Excel file.
}
\examples{
\dontrun{
import("DEMO", quarter = "23Q1")
pids_cases <- sample(Demo$primaryid, 100)
RG <- sample(Demo[!primaryid \%in\% pids_cases]$primaryid, 100)

# Generate descriptive statistics for cases
descriptive(pids_cases = pids_cases)
descriptive(pids_cases = pids_cases, drug = "paracetamol")

# Generate descriptive statistics for cases and non-cases
descriptive(pids_cases = pids_cases, RG = RG)
descriptive(pids_cases = pids_cases, RG = RG, drug = "paracetamol")
}
}
