% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiAna-package.R
\docType{package}
\name{DiAna-package}
\alias{DiAna}
\alias{DiAna-package}
\title{DiAna: DIsproportionality ANAlysis and Pharmacovigilance in the FAERS}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Advanced tools for analyzing the FDA Adverse Event Reporting System (FAERS) data, enhancing drug safety research. Provides functions to download cleaned FAERS data and import relational datasets, facilitating case identification and control selection to address confounding in analyses. Provides functions for descriptive and disproportionality analysis, time-to-onset analysis, network analysis, retrieval algorithms, and visualization. The drug name standardization is described in Fusaroli et al. (2024) \doi{10.1007/s40264-023-01391-4}, the network analysis in Fusaroli et al. (2021) \doi{10.3389/fphar.2021.740707}, the Omega interaction measure in Noren et al. (2008) \doi{10.1002/sim.3247}, and the pregnancy algorithm extends Sakai et al. (2022) \doi{10.3389/fphar.2022.1063625}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fusarolimichele.github.io/DiAna_package/}
  \item \url{https://github.com/fusarolimichele/DiAna_package}
  \item \url{https://github.com/fusarolimichele/DiAna_cleaning}
  \item \url{https://osf.io/zqu89/}
  \item Report bugs at \url{https://github.com/fusarolimichele/DiAna_package/issues}
}

}
\author{
\strong{Maintainer}: Michele Fusaroli \email{michele.fusaroli@gmail.com} (\href{https://orcid.org/0000-0002-0254-2212}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Michele Fusaroli \email{michele.fusaroli@gmail.com} (\href{https://orcid.org/0000-0002-0254-2212}{ORCID}) [copyright holder]
  \item Valentina Giunchi \email{valentina.giunchi2@unibo.it} (\href{https://orcid.org/0000-0001-5841-8520}{ORCID})
}

Other contributors:
\itemize{
  \item Violetta Zoffoli [contributor]
  \item Lionel Van Holle \email{lionel.vanholle@opensourcepv.com} (\href{https://orcid.org/0000-0001-7354-6475}{ORCID}) [contributor]
  \item Takamasa Sakai \email{tksakai@meijo-u.ac.jp} [contributor]
  \item Nhung Trinh \email{nhung.trinh@farmasi.uio.no} [contributor]
}

}
\keyword{internal}
