% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiAna-package.R
\docType{package}
\name{DiAna-package}
\alias{DiAna}
\alias{DiAna-package}
\title{DiAna: Advanced Disproportionality Analysis in the FAERS for Drug Safety}
\description{
The `DiAna` (Disproportionality Analysis) package allows for downloading FAERS cleaned data, and import relational datasets. It allows to identify cases and select controls to reduce confounding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fusarolimichele/DiAna_package}
  \item \url{https://github.com/fusarolimichele/DiAna}
  \item \url{https://osf.io/zqu89/}
  \item Report bugs at \url{https://github.com/fusarolimichele/DiAna_package/issues}
}

}
\author{
\strong{Maintainer}: Michele Fusaroli \email{michele.fusaroli2@unibo.it} (\href{https://orcid.org/0000-0002-0254-2212}{ORCID})

Authors:
\itemize{
  \item Valentina Giunchi \email{valentina.giunchi2@unibo.it} (\href{https://orcid.org/0000-0001-5841-8520}{ORCID})
  \item Lionel Van Holle \email{lionel.vanholle@opensourcepv.com} (\href{https://orcid.org/0000-0001-7354-6475}{ORCID})
}

}
\keyword{internal}
