% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANA_reference.R
\name{DiAna_reference}
\alias{DiAna_reference}
\title{Get DiAna reference}
\usage{
DiAna_reference(print = TRUE)
}
\arguments{
\item{print}{Logical, should the output be printed? Default to TRUE.}
}
\value{
DiAna reference
}
\description{
This function provides the DiAna reference, that should be used when using DiAna.
}
\examples{
\dontrun{
# DiAna_reference()
}
}
