% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{disproportionality_analisis}
\alias{disproportionality_analisis}
\title{Perform Disproportionality Analysis}
\usage{
disproportionality_analisis(
  drug_selected,
  reac_selected,
  temp_d = Drug,
  temp_r = Reac,
  meddra_level = "pt",
  restriction = "none",
  ROR_minimum_cases = 3,
  ROR_threshold = 1,
  IC_threshold = 0
)
}
\arguments{
\item{drug_selected}{A list of drugs for analysis.}

\item{reac_selected}{A list of adverse events for analysis.}

\item{temp_d}{Data table containing drug data (default is Drug).}

\item{temp_r}{Data table containing reaction data (default is Reac).}

\item{meddra_level}{The desired MedDRA level for analysis (default is "pt").}

\item{restriction}{Primary IDs to consider for analysis (default is "none",
which includes the entire population).}

\item{ROR_minimum_cases}{Threshold for calculating Reporitng Odds Ratio (default is 3).}

\item{ROR_threshold}{Threshold for analysis (default is 1).}

\item{IC_threshold}{Threshold for calculating Information Component (default is 1).}
}
\value{
A data.table containing disproportionality analysis results.
}
\description{
performs disproportionality analysis and returns the results.
}
