% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{disproportionality_analysis}
\alias{disproportionality_analysis}
\title{Perform Disproportionality Analysis}
\usage{
disproportionality_analysis(
  drug_selected,
  reac_selected,
  temp_d = Drug,
  temp_r = Reac,
  meddra_level = "pt",
  drug_level = "substance",
  restriction = "none",
  ROR_minimum_cases = 3,
  ROR_threshold = 1,
  IC_threshold = 0
)
}
\arguments{
\item{drug_selected}{A list of drugs for analysis. Can be a list of lists (to collapse terms together) if drug_level is set to custom.}

\item{reac_selected}{A list of adverse events for analysis. Can be a list of lists (to collapse terms together) if meddra_level is set to custom.}

\item{temp_d}{Data table containing drug data (default is Drug). If set to Drug[role_cod \%in\% c("PS","SS")] allows to investigate only suspects.}

\item{temp_r}{Data table containing reaction data (default is Reac).}

\item{meddra_level}{The desired MedDRA level for analysis (default is "pt"). If set to "custom" allows a list of lists for reac_selected (collapsing multiple terms).}

\item{drug_level}{The desired drug level for analysis (default is "substance"). If set to "custom" allows a list of lists for reac_selected (collapsing multiple terms).}

\item{restriction}{Primary IDs to consider for analysis (default is "none", which includes the entire population). If set to Demo[!RB_duplicates_only_susp]$primaryid, for example, allows to exclude duplicates according to one of the deduplication algorithms.}

\item{ROR_minimum_cases}{Threshold of minimum cases for calculating Reporitng Odds Ratio (default is 3).}

\item{ROR_threshold}{Threshold for defining the significance of the lower limit of the Reporting Odds Ratio (default is 1).}

\item{IC_threshold}{Threshold for defining the significance of the lower limit of the Information Component (default is 0).}
}
\value{
A data.table containing disproportionality analysis results.
}
\description{
performs disproportionality analysis and returns the results.
}
