% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{disproportionality_comparison}
\alias{disproportionality_comparison}
\title{Disproportionality Analysis using numbers}
\usage{
disproportionality_comparison(
  drug_count = length(pids_drug),
  event_count = length(pids_event),
  drug_event_count = length(intersect(pids_drug, pids_event)),
  tot = nrow(Demo)
)
}
\value{
A data.table containing disproportionality analysis results.
}
\description{
performs disproportionality analysis when provided with numbers and returns the results.
}
\examples{
\dontrun{
disproportionality_comparison(event_count = 10000, drug_count = 3000, drug_event_count = 500, tot = 20000000)
}
}
