% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{disproportionality_trend}
\alias{disproportionality_trend}
\title{Disproportionality Time Trend for a Drug-Event Combination}
\usage{
disproportionality_trend(
  drug_selected,
  reac_selected,
  temp_drug = Drug,
  temp_reac = Reac,
  temp_demo = Demo,
  temp_demo_supp = Demo_supp[, .(primaryid, quarter)],
  meddra_level = "pt",
  drug_level = "substance",
  restriction = "none",
  time_granularity = "year",
  cumulative = TRUE,
  min_2004 = TRUE
)
}
\arguments{
\item{drug_selected}{Drug selected}

\item{reac_selected}{Event selected}

\item{temp_drug}{Drug dataset. Can be set to sample_Drug for testing}

\item{temp_reac}{Reac dataset. Can be set to sample_Reac for testing}

\item{temp_demo}{Demo dataset. Defaults to Demo. Can be se to sample_Demo for testing}

\item{temp_demo_supp}{Data frame containing supplementary demographic data, and in particular the quarter. Defaults to \verb{Demo_supp\[, .(primaryid, quarter)\]}.}

\item{meddra_level}{The desired MedDRA level for analysis (default is "pt").}

\item{drug_level}{The desired drug level for analysis (default is "substance"). If set to "custom" allows a list of lists for reac_selected (collapsing multiple terms).}

\item{restriction}{Primary IDs to consider for analysis (default is "none", which includes the entire population). If set to Demo[!RB_duplicates_only_susp]$primaryid, for example, allows to exclude duplicates according to one of the deduplication algorithms.}

\item{time_granularity}{Character string specifying the time granularity. Options are "year", "quarter", or "month". Defaults to "year".}

\item{cumulative}{Logical indicating whether to calculate cumulative values. Defaults to \code{TRUE}.}

\item{min_2004}{Logical indicating whether to start the analysis only from 2004, year of the FDA AERS first implementation. Defaults to \code{TRUE}.}
}
\value{
A data frame containing the disproportionality results over time, including:
\item{period}{Time period. Deafult is 'year'. Other values are 'quarter' and 'month'. When using 'quarter' Demo_supp is required}
\item{TOT}{Total number of reports}
\item{D_E}{Number of reports with both drug and event}
\item{D_nE}{Number of reports with the drug but not the event}
\item{D}{Total number of reports with the drug}
\item{nD_E}{Number of reports with the event but not the drug}
\item{E}{Total number of reports with the event}
\item{nD_nE}{Number of reports with neither the drug nor the event}
\item{ROR_median}{Reporting odds ratio (ROR) median}
\item{ROR_lower}{ROR lower bound (2.5\%)}
\item{ROR_upper}{ROR upper bound (97.5\%)}
\item{p_value_fisher}{Fisher's exact test p-value}
\item{Bonferroni}{Bonferroni-corrected p-value}
\item{IC_median}{Information component (IC) median}
\item{IC_lower}{IC lower bound}
\item{IC_upper}{IC upper bound}
\item{label_ROR}{Formatted ROR label}
\item{label_IC}{Formatted IC label}
}
\description{
This function calculates the disproportionality time trend for a given drug-event combination.
Check \code{disproportionality_analysis()} for more options to address biases.
Its results can be visualized using \code{plot_disproportionality_trend()}.
}
\details{
The function processes the provided data to calculate the reporting odds ratio (ROR) and the information component (IC) for the specified drug-event combination over time.
}
\examples{
drug_selected <- "paracetamol"
reac_selected <- "overdose"
result <- disproportionality_trend(drug_selected, reac_selected,
  temp_drug = sample_Drug, temp_reac = sample_Reac, temp_demo = sample_Demo,
  temp_demo_supp = sample_Demo_supp[, .(primaryid, quarter)]
)
}
\seealso{
Other disproportionality functions:
\code{\link[=disproportionality_analysis]{disproportionality_analysis()}},
\code{\link[=disproportionality_comparison]{disproportionality_comparison()}}
}
\concept{disproportionality functions}
