% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extractSMQ}
\alias{extractSMQ}
\title{Extract Standardised MedDRA Queries (SMQs)}
\usage{
extractSMQ(Narrow = TRUE)
}
\arguments{
\item{Narrow}{Logical, default is \code{TRUE}. If \code{TRUE}, only the "Narrow" scope SMQs are returned.
If \code{FALSE}, all SMQs are included.}
}
\value{
A named list where each element corresponds to an SMQ group. The names represent the SMQ category,
and the values are character vectors of Preferred Terms (PTs) associated with that SMQ.
}
\description{
This function extracts and organizes Standardised MedDRA Queries (SMQs) from an external CSV file.
The function categorizes SMQs into different levels and provides either a "Narrow" scope or all available terms.
}
\details{
The function reads the SMQ dictionary from an external CSV file located at:
\code{external_sources/smq_dictionary.csv}.
Since MedDRA subscription is required, the user must obtain the SMQ dictionary separately.
Instructions for setting up the required files are provided in the DiAna GitHub repository:
\url{https://github.com/fusarolimichele/DiAna_cleaning}.

The function processes five hierarchical levels (\code{SMQ_1} to \code{SMQ_5}) and assigns Preferred Terms (PTs)
accordingly, filtering by "Narrow" scope if requested.
}
\note{
If the required SMQ dictionary file is missing, the function stops execution and returns an error message.
}
\examples{
\dontrun{
smq_list <- extractSMQ(Narrow = TRUE)
smq_list <- extractSMQ(Narrow = FALSE)
}

}
