% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANA_reference.R
\name{FAERS_quarter_specifics}
\alias{FAERS_quarter_specifics}
\title{Get specifics of dictionaries used in database version}
\usage{
FAERS_quarter_specifics(quarter = FAERS_version, print = TRUE)
}
\value{
Specifics
}
\description{
This function provides specifics on dictionaries used for cleaning the specified database version
}
\examples{
\dontrun{
# FAERS_quarter_specifics()
}
}
