% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugnames.R
\name{Fix_DiAna_dictionary_locally}
\alias{Fix_DiAna_dictionary_locally}
\title{Fix DiAna Dictionary Locally}
\usage{
Fix_DiAna_dictionary_locally(changes_xlsx_name)
}
\arguments{
\item{changes_xlsx_name}{A string specifying the name of the Excel file containing the changes.}
}
\value{
A data.table with the updated Drug information.
}
\description{
This function updates the DiAna dictionary based on changes specified in an Excel file. It imports necessary data, identifies records to be fixed, and updates the dictionary accordingly.
}
\details{
The function performs the following steps:
\itemize{
\item Reads the changes from the specified Excel file.
\item Imports the necessary data tables (\code{DRUG_NAME} and \code{DRUG}) if they do not already exist.
\item Identifies the records in \code{Drug_name} that need to be fixed based on the changes.
\item Joins the changes with the identified records and updates the \code{Drug} table.
\item Removes the old records and adds the updated records to the \code{Drug} table.
}
}
\examples{
# This example needs that DiAna dictionary is downloaded (using setup_DiAna),
# and that an excel file with intended changes is available
if (file.exist("changes.xlsx") & file.exist("external_source/DiAna_dictionary.csv")) {
  Drug <- Fix_DiAna_dictionary_locally("changes.xlsx")
}
}
