% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{hierarchy_reactions}
\alias{hierarchy_reactions}
\title{Generate Hierarchy of Reactions}
\usage{
hierarchy_reactions(pids_cases, reactions_path)
}
\arguments{
\item{pids_cases}{Vector of primary IDs identifying the sample of interest.}

\item{reactions_path}{Path to save the CSV file containing the hierarchy.}
}
\value{
None. The function generates and writes the hierarchy to the csv.
SOCs are ordered by occurrences and, within, HLGTs, HLTs, PTs.
}
\description{
This function generates a hierarchy of reactions
based on different MedDRA levels
and writes the result to a CSV file.
}
\examples{
\dontrun{
hierarchy_reactions(pids_cases, "reactions_list.csv")
}
}
