% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{hierarchycal_rates}
\alias{hierarchycal_rates}
\title{Generate Hierarchy of events or substances}
\usage{
hierarchycal_rates(
  pids_cases,
  entity = "reaction",
  file_name = paste0(project_path, "reporting_rates.xlsx"),
  drug_role = c("PS", "SS", "I", "C")
)
}
\arguments{
\item{pids_cases}{Vector of primary IDs identifying the sample of interest.}

\item{entity}{Entity investigated. It can be one of the following:
\itemize{
\item \emph{reaction};
\item \emph{indication};
\item \emph{substance}.
}}

\item{file_name}{Path to save the XLSX file containing the hierarchy.}

\item{drug_role}{If entity is substance, it is possible to specify the drug roles that should be considered}
}
\value{
An excel file with the hierarchy of interest.
For indications and reactions, SOCs are ordered by occurrences and, within, HLGTs, HLTs, PTs.
For substances, the ATC hierarchy is followed.
}
\description{
This function generates a hierarchy of reporting rates
for a specified entity and based on different MedDRA or ATC levels
and writes the result to a xlsx file.
}
\examples{
\dontrun{
# The following examples require the MedDRA and the ATC to be imported
hierarchycal_rates(pids, "reaction", "reactions_rates.xlsx")
hierarchycal_rates(pids, "indication", "indications_rates.xlsx")
hierarchycal_rates(pids, "substance", "substances_rates.xlsx")
}
}
