% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(df_name, quarter = FAERS_version, pids = NA, save_in_environment = TRUE)
}
\arguments{
\item{df_name}{Name of the data file (without the ".rds" extension).
It can be one of the following:
\itemize{
\item \emph{DRUG} =  Suspect and concomitant drugs (active ingredient);
\item \emph{REAC} =  Suspect reactions (MedDRA PT);
\item \emph{DEMO} =  Demographics and Reporting;
\item \emph{DEMO_SUPP} =  Demographics and Reporting;
\item \emph{INDI} =  Reasons for use;
\item \emph{OUTC} =  Outcomes;
\item \emph{THER} =  Drug regimen information;
\item \emph{DOSES} =  Dosage information;
\item \emph{DRUG_SUPP} =  Dechallenge, Rechallenge, route, form;
\item \emph{DRUG_NAME} =  Suspect and concomitant drugs (raw terms);
}}

\item{quarter}{The quarter from which to import the data.
For updated analyses use last quarterly update,
in the format \emph{23Q1}.}

\item{pids}{Optional vector of primary IDs to subset the imported data.
Defaults to the entire population.}
}
\value{
A data.table containing the imported data.
}
\description{
Imports the specified FAERS relational database.
}
\examples{
\dontrun{
# Import full DRUG dataset up to 23Q1
import("DRUG", "23Q1")

# Import reac data for specific primary IDs in quarters up to 23Q1
import("REAC", "23Q1", pids = c("pid1", "pid2"))
}

}
