% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_MedDRA}
\alias{import_MedDRA}
\title{Import MedDRA Data}
\usage{
import_MedDRA(env = .GlobalEnv)
}
\arguments{
\item{env}{The environment where the data will be assigned. Default to .GlobalEnv}
}
\value{
A data table containing MedDRA data.
}
\description{
This function imports MedDRA (Medical Dictionary for Regulatory Activities) data from a CSV file and stores it in the global environment.
}
\details{
This function reads MedDRA data from a CSV file located at the path specified by \code{here()/external_sources/meddra_primary.csv}.
If the file does not exist, it will stop execution and provide instructions on how to obtain MedDRA data.
If the file exists, it will load the data, select specific columns (def, soc, hlgt, hlt, pt), remove duplicates, and store it in the global environment as "MedDRA".
}
\examples{
# This example requires a specific file that can only be available with a MeDRA subscription.
if (file.exist("external_source/meddra_primary.csv")) {
  import_MedDRA()
}

}
\seealso{
You can find more information and instructions for obtaining MedDRA data at https://github.com/fusarolimichele/DiAna.
}
