% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{network_analysis}
\alias{network_analysis}
\title{Perform Network Analysis and Visualization}
\usage{
network_analysis(
  pids,
  entity = "reaction",
  remove_singlet = TRUE,
  remove_negative_edges = TRUE,
  file_name = paste0(project_path, "network.tiff"),
  width = 1500,
  height = 1500,
  labs_size = 1,
  restriction = NA
)
}
\arguments{
\item{pids}{Numeric vector of unique identifiers for data import.}

\item{entity}{Character specifying the type of entity to analyze ("reaction", "indication", or "substance").}

\item{remove_singlet}{Logical indicating whether to remove singleton nodes (nodes with no edges). Default is TRUE.}

\item{remove_negative_edges}{Logical indicating whether to remove edges with negative weights. Default is TRUE.}

\item{file_name}{Character string specifying the file name (including path) to save the network visualization. Default is "network.tiff".}

\item{width}{Numeric specifying the width of the saved image in pixels. Default is 1500.}

\item{height}{Numeric specifying the height of the saved image in pixels. Default is 1500.}

\item{labs_size}{Size of labels in network visualization. Default is 1. It can be changed if visualization is not good.}

\item{restriction}{Restriction performed in the analysis. Default is NA. It could be setted to 'suspects' if entity is 'substance' to restrict the analysis to primary and secondary suspects}
}
\value{
NULL (invisibly). Saves a network visualization as a TIFF file.
}
\description{
This function performs network analysis on the provided data and generates a visualization of the network. It uses IsingFit to model the network structure, clusters nodes using Louvain method, and visualizes the resulting network graph.
}
\examples{
\dontrun{
# Example usage:
# Perform network analysis for reactions with specified pids
network_analysis(pids = c(1, 2, 3), entity = "reaction")
}

}
\references{
Fusaroli M, Raschi E, Gatti M, De Ponti F, Poluzzi E. Development of a Network-Based Signal Detection Tool: The COVID-19 Adversome in the FDA Adverse Event Reporting System. Front Pharmacol. 2021 Dec 8;12:740707. doi: 10.3389/fphar.2021.740707. PMID: 34955821; PMCID: PMC8694570.

Fusaroli M, Polizzi S, Menestrina L, Giunchi V, Pellegrini L, Raschi E, et al. Unveiling the Hidden Toll of Drug-Induced Impulsivity: A Network Analysis of the FDA Adverse Event Reporting System. 2023; : 2023.11.17.23298635. [Cited 2023 Nov 29] Available from https://www.medrxiv.org/content/10.1101/2023.11.17.23298635v1
}
