% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{new_descriptive}
\alias{new_descriptive}
\title{Generate Descriptive Statistics for a Sample}
\usage{
new_descriptive(
  pids_cases,
  RG = NULL,
  drug = NULL,
  save_in_excel = TRUE,
  file_name = "Descriptives.xlsx",
  vars = c("sex", "Submission", "Reporter", "age_range", "Outcome", "country",
    "continent", "age_in_years", "wt_in_kgs", "Reactions", "Indications", "Substances",
    "num_Substances", "year", "role_cod", "time_to_onset"),
  list_pids = list(),
  method = "independence_test",
  database = "sample"
)
}
\arguments{
\item{pids_cases}{A vector of primary IDs for cases.}

\item{RG}{A vector of primary IDs: reference group. Default is NULL.}

\item{drug}{A vector of drug names. Default is NULL.}

\item{save_in_excel}{Whether to save the outcome in an excel. Defaults to TRUE}

\item{file_name}{The name of the Excel file to save the results. Default is "Descriptives.xlsx". It only works if save_in_excel is TRUE.}

\item{vars}{A character vector of variable names to include in the analysis.}

\item{list_pids}{A list of vectors with primary IDs for custom groups whose distribution should be described. Default is an empty list.}

\item{method}{The method for Chi-square test analysis, either "independence_test" or "goodness_of_fit". Default is "independence_test". It applies only for comparisons between cases and non-cases.}

\item{database}{Database on which to run the analysis. By default "sample".}
}
\value{
The function generates descriptive statistics as a gt_table and potentially saves them to an Excel file.
}
\description{
This function generates descriptive statistics for a sample of data, including cases and non-cases, and saves the results to an Excel file.
}
\examples{
pids_cases <- unique(sample_Demo[sex == "M"]$primaryid)
RG <- unique(sample_Demo[sex == "M"]$primaryid)

# Generate descriptive statistics for cases
new_descriptive(
  pids_cases = pids_cases, save_in_excel = FALSE
)

# Generate descriptive statistics for cases and non-cases
new_descriptive(
  pids_cases = pids_cases, RG = RG, save_in_excel = FALSE
)
}
