% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{plot_disproportionality_trend}
\alias{plot_disproportionality_trend}
\title{Plot Disproportionality Trend}
\usage{
plot_disproportionality_trend(disproportionality_trend_results, metric = "IC")
}
\arguments{
\item{disproportionality_trend_results}{Data frame containing the results from the \code{disproportionality_trend} function.}

\item{metric}{Character string specifying the metric to plot. Options are "IC" (information component) or "ROR" (reporting odds ratio). Defaults to "IC".}
}
\value{
A ggplot object representing the disproportionality trend plot for the specified metric.
}
\description{
This function plots the disproportionality trend over time for a given metric.
}
\details{
The function creates a plot to visualize the disproportionality trend of a drug-event combination over time. Depending on the selected metric, it plots either the information component (IC) or the reporting odds ratio (ROR) with corresponding confidence intervals.
}
\examples{
\dontrun{
# Example usage
trend_results <- disproportionality_trend(drug_selected = "aspirin", reac_selected = "headache")
plot_IC <- plot_disproportionality_trend(trend_results, metric = "IC")
plot_ROR <- plot_disproportionality_trend(trend_results, metric = "ROR")
}

}
