% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{render_forest}
\alias{render_forest}
\title{Render Forest Plot}
\usage{
render_forest(
  df,
  TTO_ROR = "ROR",
  levs_row = unique(df$substance),
  facet_v = "event",
  facet_h = NA,
  row = "substance",
  nested = FALSE,
  text_size_legend = 15,
  transformation = "identity",
  nested_colors = NA,
  threshold = 1,
  dodge = 0.3,
  show_legend = FALSE
)
}
\arguments{
\item{df}{Data.table containing the data for rendering the forest plot.}

\item{TTO_ROR}{Type of data to use for rendering: "TTO", "ROR", "ROR_disp", or "IC".}

\item{levs_row}{Levels for the rows of the forest plot.}

\item{facet_v}{Variable for vertical facetting (default is "event").}

\item{facet_h}{Variable for horizontal facetting (default is NA).}

\item{row}{Variable for the rows of the forest plot (default is "drug").}

\item{nested}{Logical indicating if nested plotting is required (default is FALSE).}

\item{text_size_legend}{Size of text in the legend (default is 15).}

\item{transformation}{Transformation for the x-axis (default is "identity").}

\item{nested_colors}{Vector of colors for plot elements.}

\item{threshold}{Threshold value for analysis (default is 1).}

\item{dodge}{Position adjustment for dodging (default is 0.3).}

\item{show_legend}{Logical indicating whether to show the legend (default is FALSE).}
}
\value{
A ggplot object representing the forest plot visualization.
}
\description{
This function generates a forest plot visualization of disproportions.
}
