% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality_analysis.R
\name{render_forest}
\alias{render_forest}
\title{Render Forest Plot}
\usage{
render_forest(
  df,
  index = "IC",
  row = "substance",
  levs_row = NA,
  nested = FALSE,
  show_legend = TRUE,
  transformation = "identity",
  custom_threshold = NA,
  text_size_legend = 15,
  dodge = 0.3,
  nested_colors = NA,
  facet_v = NA,
  facet_h = NA
)
}
\arguments{
\item{df}{Data.table containing the data for rendering the forest plot.}

\item{index}{Type of data to use for rendering: "ROR", "ROR_disp", or "IC".}

\item{row}{Variable for the rows of the forest plot (default is "drug").}

\item{levs_row}{Levels for the rows of the forest plot.}

\item{nested}{Variable indicating if nested plotting is required (default is FALSE). If nested plotting is required the name of the variable should replace FALSE.}

\item{show_legend}{Logical indicating whether to show the legend (default is FALSE).}

\item{transformation}{Transformation for the x-axis (default is "identity").}

\item{custom_threshold}{Threshold value for analysis (default is 1).}

\item{text_size_legend}{Size of text in the legend (default is 15).}

\item{dodge}{Position adjustment for dodging (default is 0.3).}

\item{nested_colors}{Vector of colors for plot elements.}

\item{facet_v}{Variable for vertical facetting (default is "event").}

\item{facet_h}{Variable for horizontal facetting (default is NA).}
}
\value{
A ggplot object representing the forest plot visualization.
}
\description{
This function generates a forest plot visualization of disproportions.
}
