% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{reporting_rates}
\alias{reporting_rates}
\title{Reporting rates of events or substances}
\usage{
reporting_rates(
  pids_cases,
  entity = "reaction",
  level = "pt",
  drug_role = c("PS", "SS", "I", "C"),
  drug_indi = NA
)
}
\arguments{
\item{pids_cases}{Vector of primary IDs to consider for counting reactions.}

\item{entity}{Entity investigated. It can be one of the following:
\itemize{
\item \emph{reaction};
\item \emph{indication};
\item \emph{substance}.
}}

\item{level}{The desired MedDRA or ATC level for counting (default is "pt").}

\item{drug_role}{is only used for substances. By default both suspect and concomitant drugs are included.}

\item{drug_indi}{is only used for indications. By default the indications of all the drugs of the selected primaryids are considered, but you can specify a vector of drugs.}
}
\value{
A data.table containing counts and percentages of the investigated entity
at the specified level and in descending order.
}
\description{
counts the occurrences of reactions/indications/substance
for a given set of primary IDs.
Calculates \% as the proportion of individuals recording the event.
}
\examples{
\dontrun{
reporting_rates(pids, "reaction", "hlt")
reporting_rates(pids, "indication", "pt")
reporting_rates(pids, entity = "substance", level = "Class3")
}
}
