% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reactions.R
\name{reporting_rates}
\alias{reporting_rates}
\title{Reporting rates of events or substances}
\usage{
reporting_rates(
  pids_cases,
  entity = "reaction",
  level = "pt",
  drug_role = c("PS", "SS", "I", "C"),
  drug_indi = NA,
  temp_reac = Reac,
  temp_drug = Drug,
  temp_indi = Indi
)
}
\arguments{
\item{pids_cases}{Vector of primary IDs to consider for counting reactions.}

\item{entity}{Entity investigated. It can be one of the following:
\itemize{
\item \emph{reaction};
\item \emph{indication};
\item \emph{substance}.
}}

\item{level}{The desired MedDRA or ATC level for counting (default is "pt").}

\item{drug_role}{is only used for substances. By default both suspect and concomitant drugs are included.}

\item{drug_indi}{is only used for indications. By default the indications of all the drugs of the selected primaryids are considered, but you can specify a vector of drugs.}

\item{temp_reac}{Reac dataset. Can be set to sample_Reac for testing}

\item{temp_drug}{Drug dataset. Can be set to sample_Drug for testing}

\item{temp_indi}{Indi dataset. Can be set to sample_Indi for testing}
}
\value{
A data.table containing counts and percentages of the investigated entity
at the specified level and in descending order.
}
\description{
counts the occurrences of reactions/indications/substance
for a given set of primary IDs.
Calculates \% as the proportion of individuals recording the event.
}
\examples{
# select only reports recording paracetamol from the sample dataset,
# and provide the most reported events (at the pt level),
# drugs, and indications (at the PT level).
pids_cases <-
reporting_rates(unique(sample_Drug[substance == "paracetamol"]$primaryid),
 "reaction", "pt",
  temp_reac = sample_Reac,
  temp_drug = sample_Drug, temp_indi = sample_Indi
)
reporting_rates(unique(sample_Drug[substance == "paracetamol"]$primaryid),
  "indication", "pt",
  temp_reac = sample_Reac,
  temp_drug = sample_Drug, temp_indi = sample_Indi
)
reporting_rates(unique(sample_Drug[substance == "paracetamol"]$primaryid),
  entity = "substance",
  temp_reac = sample_Reac, temp_drug = sample_Drug,
  temp_indi = sample_Indi
)
}
