% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cases.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve data for specific primaryids}
\usage{
retrieve(
  pids,
  file_name = "individual_cases",
  temp_reac = Reac,
  temp_drug = Drug,
  temp_demo = Demo,
  temp_demo_supp = Demo_supp,
  temp_outc = Outc,
  temp_ther = Ther,
  temp_doses = Doses,
  temp_drug_supp = Drug_supp,
  temp_indi = Indi,
  temp_drug_name = Drug_name,
  temp_meddra = NA,
  temp_atc = NA,
  save_in_excel = !missing(file_name)
)
}
\arguments{
\item{pids}{Primaryids of interest.}

\item{file_name}{The name of the output xlsx file (default is "individual cases").}

\item{temp_reac}{Reac dataset. Can be set to sample_Reac for testing}

\item{temp_drug}{Drug dataset. Can be set to sample_Drug for testing}

\item{temp_demo}{Demo dataset. Defaults to Demo. Can be se to sample_Demo for testing}

\item{temp_demo_supp}{the Demo_supp databases. Can be set to sample_Demo_Supp for testing}

\item{temp_outc}{Outc dataset. Can be set to sample_Outc for testing}

\item{temp_ther}{Ther dataset. Can be set to sample_Ther for testing}

\item{temp_doses}{the Doses databases. Can be set to sample_Doses for testing}

\item{temp_drug_supp}{the Drug_supp databases. Can be set to sample_Drug_Supp for testing}

\item{temp_indi}{Indi dataset. Can be set to sample_Indi for testing}

\item{temp_drug_name}{the Drug_name databases. Can be set to sample_Drug_Name for testing}

\item{temp_meddra}{the MedDRA dictionary (a data.table with columns soc, hlgt and pt), if available. Defaults to NA because MedDRA requires a subscription.}

\item{temp_atc}{the ATC classification, as returned by import_ATC(), if available. Defaults to NA. Only the primary ATC code of each substance is used.}

\item{save_in_excel}{Whether to also save the results in an Excel file, \code{file_name}. Defaults to \code{TRUE} only if \code{file_name} is supplied.}
}
\value{
A list of two data.tables with individual cases information:
\code{general_info}, with a row per ICSR,
and \code{drug_info}, with drug information and multiple rows per ICSR.
If \code{save_in_excel = TRUE}, they are also saved as \code{file_name.xlsx}
and \code{file_name_drug.xlsx}.
}
\description{
This function retrieves data for a specific group of primaryids,
allowing for in-depth case-by-case evaluation and clinical reasoning.
If MedDRA is available PTs are clustered by HLGT.
}
\examples{
pids <- unique(sample_Demo[sex == "M"]$primaryid)
cases <- retrieve(pids,
  temp_reac = sample_Reac, temp_drug = sample_Drug, temp_demo = sample_Demo,
  temp_demo_supp = sample_Demo_Supp, temp_outc = sample_Outc,
  temp_ther = sample_Ther, temp_doses = sample_Doses,
  temp_drug_supp = sample_Drug_Supp, temp_indi = sample_Indi,
  temp_drug_name = sample_Drug_Name
)
head(cases$general_info)

}
