% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cases.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve data for specific primaryids}
\usage{
retrieve(pids, file_name = "individual_cases", quarter = FAERS_version)
}
\arguments{
\item{pids}{Primaryids of interest.}

\item{file_name}{The name of the output xlsx file (default is "individual cases").}

\item{quarter}{The quarter for data retrieval (default is FAERS_version).}
}
\value{
Two xlsx files with individual cases information:
one general with a row per ICSR,
and one with drug information and multiple rows per ICSR.
}
\description{
This function retrieves data for a specific group of primaryids,
allowing for in-depth case-by-case evaluation and clinical reasoning.
If MedDRA is available PTs are clustered by HLGT.
}
\examples{
\dontrun{
retrieve(c(1, 2, 3), "output_data")
}

}
