% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy_algorithm.R
\name{retrieve_pregnancy_pids}
\alias{retrieve_pregnancy_pids}
\title{Retrieve Pregnancy-Related Report Identifiers from FAERS}
\usage{
retrieve_pregnancy_pids(quarter = FAERS_version)
}
\arguments{
\item{quarter}{A character string specifying the FAERS quarter to retrieve data for. Default is the current FAERS version (\code{FAERS_version}). It can also be set to "sample" to use instead sample data, mainly for testing.}
}
\value{
A list containing five elements:
\describe{
\item{\code{high_specificity}}{A vector of identifiers with high specificity for pregnancy-related reports.}
\item{\code{medium_specificity}}{A vector of identifiers with medium specificity for pregnancy-related reports.}
\item{\code{low_specificity}}{A vector of identifiers with low specificity for pregnancy-related reports.}
\item{\code{paternal_exposure}}{A vector of identifiers for paternal exposure-related reports.}
\item{\code{flowchart}}{An editable flowchart showing case retrieval}
}
}
\description{
This function retrieves the identifiers of pregnancy-related reports from the FDA Adverse Event Reporting System (FAERS) for a specified quarter.
}
\details{
The function processes data from multiple FAERS tables (\code{DEMO}, \code{DRUG}, \code{REAC}, \code{INDI}, \code{OUTC}, \code{THER}, and \code{DRUG_SUPP}) to identify pregnancy-related reports based on specific indications, reactions, and drug routes. The results are filtered to exclude reports unlikely to be related to pregnancy (e.g., reports involving males, children, or older adults). The algorithm is an implementation and evolution of the original pregnancy algorithm by Sakai,ref. 10.3389/fphar.2022.1063625
#' @references
Sakai T, Mori C, Ohtsu F. Potential safety signal of pregnancy loss with vascular endothelial growth factor inhibitor intraocular injection: A disproportionality analysis using the Food and Drug Administration Adverse Event Reporting System. Front Pharmacol. 2022 Nov 10;13:1063625. doi: 10.3389/fphar.2022.1063625. PMID: 36438807; PMCID: PMC9684212.
}
\examples{
# On the sample data shipped with DiAna
pids_pregnancy <- retrieve_pregnancy_pids(quarter = "sample")
pids_pregnancy$medium_specificity

# On the entire database, which requires the data downloaded with setup_DiAna()
if (file.exists(file.path(here::here(), "data", "24Q1", "DEMO.rds"))) {
  pids_pregnancy <- retrieve_pregnancy_pids(quarter = "24Q1")
}
}
