% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Demo}
\alias{sample_Demo}
\title{Sample Demographic Data}
\format{
A data frame with 15 variables:
\describe{
\item{primaryid}{Unique identifier for each individual (\code{numeric}).}
\item{sex}{Sex of the individual (\code{factor}), with levels \code{M} for male and \code{F} for female.}
\item{age_in_days}{Age of the individual in days (\code{numeric}).}
\item{wt_in_kgs}{Weight of the individual in kilograms (\code{numeric}).}
\item{occr_country}{Country where the event occurred (\code{factor}).}
\item{event_dt}{Date of the event (\code{integer}), in the format YYYYMMDD.}
\item{occp_cod}{Occupation code of the individual (\code{factor}).}
\item{reporter_country}{Country of the reporter (\code{factor}).}
\item{rept_cod}{Report code (\code{factor}).}
\item{init_fda_dt}{Initial FDA date (\code{integer}), in the format YYYYMMDD.}
\item{fda_dt}{FDA date (\code{integer}), in the format YYYYMMDD.}
\item{premarketing}{Logical indicator if the case is premarketing (\code{logical}).}
\item{literature}{Logical indicator if the case is from literature (\code{logical}).}
\item{RB_duplicates}{Logical indicator if the case is a duplicate based on the Rule Based algorithm (\code{logical}).}
\item{RB_duplicates_only_susp}{Logical indicator if the case is a duplicate based on the Rule Based algorithm considering only suspected drugs (\code{logical}).}
}
}
\source{
https://github.com/fusarolimichele/DiAna
}
\usage{
sample_Demo
}
\description{
A dataset containing demographic information of individuals with various details
such as sex, age in days, weight in kilograms, and other related
information.
}
\details{
This dataset is a subset of the actual Demo data (that can be downloaded using setup_DiAna()),
providing a glimpse into the structure of the database and allowing tests and documentation.
}
\keyword{datasets}
