% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Demo_Supp}
\alias{sample_Demo_Supp}
\title{Sample Supplementary Demographic Data}
\format{
A data.table with 10 variables:
\describe{
\item{primaryid}{Unique identifier for each individual (\code{numeric}).}
\item{rpsr_cod}{Report source code (\code{factor}).}
\item{caseid}{Case ID to link the report with the public dashboard (\code{integer}).}
\item{caseversion}{Case version in case of multiple updates (\code{factor}).}
\item{i_f_cod}{Initial or follow-up version (\code{factor}).}
\item{auth_num}{Authorization number (\code{character}).}
\item{e_sub}{Electronic submission  (\code{factor}).}
\item{lit_ref}{Literature reference for the case (\code{character}).}
\item{rept_dt}{Reporting date (\code{integer}).}
\item{to_mfr}{Sent to manufacturer (\code{character}).}
\item{mfr_sndr}{Manufacturer sender (\code{character}).}
\item{mfr_num}{Manufacturer id number (\code{character}).}
\item{mfr_dt}{Manufacturer date (\code{integer}).}
\item{quarter}{FAERS Quarter in which the report is stored  (\code{factor}).}
}
}
\source{
https://github.com/fusarolimichele/DiAna
}
\usage{
sample_Demo_Supp
}
\description{
A dataset containing supplementary demographic information of individuals,
including reporter codes, case IDs, and additional details.
}
\details{
This subset of Demo_Supp provides supplementary demographic information of reports,
including details related to reporting and manufacturer information.
It is useful for understanding additional context around adverse drug reaction reports.
}
\keyword{datasets}
