% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Doses}
\alias{sample_Doses}
\title{Sample Data on Drug Doses}
\format{
A data.table with 7 variables:
\describe{
\item{primaryid}{Unique identifier for each individual (\code{numeric}).}
\item{drug_seq}{Sequence number identifying the specific drug in the report (\code{integer}).}
\item{dose_vbm}{Verbatim (free-text) dose description (\code{character}).}
\item{dose_unit}{Unit of dose (\code{character}).}
\item{cum_dose_chr}{Cumulative dose as character (\code{numeric}).}
\item{dose_amt}{Amount of dose (\code{character}).}
\item{cum_dose_unit}{Cumulative dose unit (\code{character}).}
\item{dose_freq}{Frequency of dose (\code{factor}).}
}
}
\source{
https://github.com/fusarolimichele/DiAna_cleaning
}
\usage{
sample_Doses
}
\description{
A dataset containing information on drug doses administered,
linking primary IDs with dose details.
}
\details{
This subset of Doses provides detailed information on drug doses administered,
including verbatim descriptions, units, cumulative doses, amounts,
and frequency of administration, facilitating analysis related to dosing patterns.
As it can be observed in the sample, most of the dataset is NA (info not available).
}
\keyword{datasets}
