% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Indi}
\alias{sample_Indi}
\title{Sample dataset containing indications for using reported drugs.}
\format{
A data.table with 3 variables:
\describe{
\item{primaryid}{Unique identifier for each report.}
\item{drug_seq}{Sequence number identifying a specific drug reported in a report.}
\item{indi_pt}{Preferred term describing the drug indication according to the Medical Dictionary for Regulatory Activities.}
}
}
\source{
https://github.com/fusarolimichele/DiAna
}
\usage{
sample_Indi
}
\description{
This dataset provides information on indications.
Each report (primaryid) can record multiple drugs and therefore multiple indications (and rows).
}
\details{
This dataset provides detailed information on drug indications reported for individuals,
categorized by primary ID and drug sequence, facilitating analysis related to specific
indications observed in adverse drug reaction reports. Each report can record multiple drugs
and therefore multiple indications.
}
\keyword{datasets}
