% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Outc}
\alias{sample_Outc}
\title{Sample Data on Individual Outcomes}
\format{
A data.table with 2 variables:
\describe{
\item{primaryid}{Unique identifier for each individual (\code{numeric}).}
\item{outc_cod}{Outcome code indicating the severity or outcome of the adverse event (\code{ordered}).}
}
}
\source{
https://github.com/fusarolimichele/DiAna
}
\usage{
sample_Outc
}
\description{
A dataset containing individual-level outcomes associated with adverse drug reactions,
linking primary IDs with outcome codes.
}
\details{
This dataset provides information on outcomes associated with reported adverse drug reactions,
specifying the severity or nature of each event observed for individuals.
}
\keyword{datasets}
