% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_DIANA.R
\name{setup_DiAna}
\alias{setup_DiAna}
\title{Set Up DiAna Environment}
\usage{
setup_DiAna(quarter = "23Q1")
}
\arguments{
\item{quarter}{The quarter for which to set up the DiAna environment
(default is "23Q1"). The ones available:
\itemize{
\item \emph{23Q1}
}}
}
\value{
None. The function sets up the environment and downloads data.
}
\description{
This function sets up the DiAna environment by creating necessary folders
and downloading the DiAna data up to the specified quarter.
}
\examples{
\dontrun{
# Set up DiAna environment for the default quarter
setup_DiAna()
}
}
