% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_DIANA.R
\name{setup_DiAna}
\alias{setup_DiAna}
\title{setup_DiAna}
\usage{
setup_DiAna(quarter = "23Q1")
}
\arguments{
\item{quarter}{The ones available:
\itemize{
\item \emph{23Q1}
}}
}
\value{
It creates, on the desktop, the directory which will host all the disproportionality analyses. It also download the cleaned FAERS up to the quarter of year specified.
}
\description{
Sets up DiAna on the desktop, downloading the cleaned FAERS up to the specified quarter.
}
\examples{
\dontrun{
setup_DiAna()
}
}
