% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snippets.R
\name{snippets_install_github}
\alias{snippets_install_github}
\title{Install RStudio Snippets from GitHub Repository}
\usage{
snippets_install_github(repo = "fusarolimichele/DiAna_snippets")
}
\arguments{
\item{repo}{Character. The GitHub repository containing the snippets. Default is "fusarolimichele/DiAna_snippets".}
}
\value{
None. This function is called for its side effects, which include installing RStudio snippets.
}
\description{
This function installs RStudio code snippets from a specified GitHub repository.
}
\examples{
\dontrun{
# This example is using internet connection to download snippets to initialize scripts.
# It automatically includes snippets among the ones available to the user.
# It should not be run at the check.
snippets_install_github()
}
}
