% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_onset_analysis.R
\name{time_to_onset_analysis}
\alias{time_to_onset_analysis}
\title{Perform time-to-onset analysis for drug-event associations}
\usage{
time_to_onset_analysis(
  drug_selected,
  reac_selected,
  temp_d = Drug,
  temp_r = Reac,
  meddra_level = "pt",
  drug_level = "substance",
  restriction = "none",
  minimum_cases = 3,
  max_TTO = 365
)
}
\arguments{
\item{drug_selected}{A list of drugs for analysis. Can be a list of lists (to collapse terms together) if drug_level is set to custom.}

\item{reac_selected}{A list of adverse events for analysis. Can be a list of lists (to collapse terms together) if meddra_level is set to custom.}

\item{temp_d}{Data table containing drug data (default is Drug). If set to Drug\link{role_cod \%in\% c("PS","SS")} allows to investigate only suspects.}

\item{temp_r}{Data table containing reaction data (default is Reac).}

\item{meddra_level}{The desired MedDRA level for analysis (default is "pt"). If set to "custom" allows a list of lists for reac_selected (collapsing multiple terms).}

\item{drug_level}{The desired drug level for analysis (default is "substance"). If set to "custom" allows a list of lists for reac_selected (collapsing multiple terms).}

\item{restriction}{Primary IDs to consider for analysis (default is "none", which includes the entire population). If set to Demo\link{!RB_duplicates_only_susp}$primaryid, for example, allows to exclude duplicates according to one of the deduplication algorithms.}

\item{minimum_cases}{The minimum number of cases required for the analysis (default is 3).}

\item{max_TTO}{The maximum time to onset considered in the analysis, in days (default is 365).}
}
\value{
A data.table containing results of the time-to-onset analysis, including drug-event associations, KS test statistics, and p-values.
}
\description{
This function conducts a time-to-onset analysis for associations between drugs and events based on user-specified parameters.
}
\examples{
\dontrun{
# Example usage:
df <- time_to_onset_analysis("aripiprazole", "gambling")
df1 <- time_to_onset_analysis(c("aripiprazole", "pramipexole"), c("gambling", "gambling disorder"))
df2 <- time_to_onset_analysis(c("aripiprazole", "pramipexole"),
  list(
    "gambling_PT" = c("gambling"),
    "gambling_disorder_PT" = c("gambling disorder"),
    "gambling_query" = c("gambling", "gambling disorder")
  ),
  temp_d = Drug[role_cod \%in\% c("PS", "SS")],
  meddra_level = "custom",
  max_TTO = 3 * 365
)
}

}
\seealso{
\code{\link{ks.test}} for information about the Kolmogorov-Smirnov test.

#' @references
Van Holle, L., Zeinoun, Z., Bauchau, V. and Verstraeten, T. (2012), Using time-to-onset for detecting safety signals in spontaneous reports of adverse events following immunization: a proof of concept study. Pharmacoepidemiol Drug Saf, 21: 603-610. https://doi.org/10.1002/pds.3226
}
